refactor: Delete duplicate ast lib
